refactor(LayerPanel): split into LayerRow, useLayerDnd, and LayerContextMenu (#461)
Extract the 220-line mega-map body into a LayerRow component, the
drag-and-drop state machine into a useLayerDnd hook, and the context
menu into a LayerContextMenu component. LayerPanel.tsx drops from
668 to 248 lines.
